Make TotemicEventHooks an interface and move implementation
