+++ src/main/java/pokefenn/totemic/PlatformAbstractions.java
@@ -4,5 +4,5 @@ package pokefenn.totemic;
  * Provides access to functions that are implemented differently for each mod loader.
  */
 public interface PlatformAbstractions {
-
+    TotemicEventHooks events();
 }
